[region-isolation] Fix code where we were not properly mapping a type into a SILFunction's context to actually do so.
I did this by adding a SILIsolationInfo::isNonSendableType API that takes
CanTypes and use it to avoid needing to lower certain CanTypes before performing
the query. I also made sure to use the partial apply callee function to map the
type into context, so we perform all of the correct substitutions.

rdar://138667211