utils: use CMake to build swift-inspect for Windows
This switches us to use CMake to build swift-inspect. In doing so, we
can now cross-compile this tool. Due to the use of CMake, we now have
install rules for the products, which allows the toolchain deployment
and simpler rules in the installer.

(cherry picked from commit c72246cb1596ebbcf783960b94357d42666b3e84)