[build-script] Adding buildbot_linux,bootstrap
Adding a preset for running a bootstrapped buildbot linux configuration
to verify that bootstrapping on environments that do not have a compiler
will still work.

Swift Syntax and the Swift Driver must be disabled since build-script
cannot schedule them at the appropriate times during the compiler build.
Since we do not start with an existing Swift compiler, they can't be
pre-built.