[Observation] include ThreadingError.cpp in build
On Linux at least, using observation can lead to an undefined reference to
swift::threading::fatal(char const*, ...) at link time. That this is not an
error on Darwin may be due to different symbol export policies.

Mirroring libswift_Concurrency, ensure that ThreadingError.cpp is present as a
local symbol in libswiftObservation.

Fixes: #75760