DiagnosticEngine: Add `DiagGroupID` to `Diagnostic`
This change addresses the following issue: when an error is being
wrapped in a warning, the diagnostic message will use the wrapper's
`DiagGroupID` as the warning's name. However, we want to retain the
original error's group for use. For example, in Swift 5,
`async_unavailable_decl` is wrapped in `error_in_future_swift_version`.
When we print a diagnostic of this kind, we want to keep the
`DiagGroupID` of `async_unavailable_decl`, not that of
`error_in_future_swift_version`.

To achieve this, we add `DiagGroupID` to the `Diagnostic` class. When an
active diagnostic is wrapped in `DiagnosticEngine`, we retain the
original `DiagGroupID`.