AST: Teach the existential generalizer to handle a protocol type's parent
The archetype-based existential erasure algorithm we transitioned to
relies on `ArchetypeType::getExistentialType` instead of
`GenericSignatureImpl::getExistentialType`. The former applies outer
substitutions computed by the existential generalizer in addition to
computing the existential upper bound. This, combined with the now
eager simplification of the type of the existential member reference
required for the archetype-based algorithm, resurrected
`validation-test/compiler_crashers_fixed/00933-no-stacktrace.swift`,
where the member access is performed on an existential of a protocol
nested in a generic context.

The existential generalizer was not producing substituions for the
protocol type's bound generic parent, which caused
`getTypeOfMemberReference` to produce a type containing type
parameters and an ensuing assertion failure.