[Backtracing] Added BacktraceBuffer.
`BacktraceBuffer` is SPI to allow a client to repeatedly capture backtraces
without allocating memory or locking.  Specifically, the backtraces are
captured directly into a ring buffer, in CBF format, and `BacktraceBuffer`
will call out to a user-provided closure whenever the buffer is full in
order to free up space.

This means the user can organise for backtrace data to be saved or transmitted
over the network or some such in large chunks, which is far more efficient
than writing individual bytes.

I've also added SPI to `Backtrace` to make it easy to decode the resulting
data stream.

rdar://117989155