[SGC] Complete unreachable-reachable lifetimes.
For a function to have complete lifetimes, the lifetime of every def in
the function which is reachable by walking backwards from an unreachable
instruction must be completed.

Currently, every def in every block which appears in the post-order of
the function's blocks is completed.  This is both too much (defs not
backwards reachable from an `unreachable` needn't be completed) and too
little (even defs in "unreachable blocks"--i.e. those which can't be
reached by a forwards walk from function entry--must be completed).

rdar://141197164