Fix Mem2Reg for enum stack locations
We complete lifetimes of enums in mem2reg since address forms may be incomplete
on non payloaded paths and promoting them to value form can cause ownership verification errors.
This was done per stack allocation. This is an issue since same value maybe stored on multiple
stack locations and can use insertion of the lifetime end on blocks where the lifetime was
previously completed by the utility.

Move lifetime completion to the end of mem2reg to not run into this.